Keep option to use config files for custom components
